requirements.txt
altgraph==0.17
gobject==0.1.0
keyboard==0.13.5
mouse==0.7.1
MouseInfo==0.1.3
Pillow==8.1.2
PyAutoGUI==0.9.52
pycairo==1.20.0
PyGetWindow==0.0.9
PyGObject==3.40.0
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
PyMsgBox==1.0.9
pyperclip==1.8.2
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
PyQtWebEngine==5.15.4
PyQtWebEngine-Qt5==5.15.2
PyRect==0.1.4
PyScreeze==0.1.26
python-xlib==0.29
python3-xlib==0.15
PyTweening==1.0.3
pywebview==3.4
six==1.15.0
